Connect to Discord gateway
Implemented a method to connect to the Discord gateway. This method
initialises a WebSocket connection to the Discord gateway return by the
HTTP call. This method sets the Socket and Heartbeat fields of the Bot instance.

Also impletemented a method that is started in a goroutine that keeps
the websocket gateway connection alive by sending a heartbeat opcode (1)
to the gateway every Bot.Heartbeat milliseconds. This implementation is
not fully operational yet.