Send identify payload and read messages
Added code to send the identify payload to the Discord gateway, hence, making
the handshake with it.

Implemented the read messages method, it will read incoming messages on
the WebSocket connection and send them to a newly implemented OnMessage
callback in the Bot type.

Since we are making the handshake with the gateway, the heartbeat
function is now working properly. Though, it should check that it
receives the Heartbeat ACK opcode from the gateway. Closing the
connection if not.