Change WebSocket dependency
Changed from the x/net/websocket library to nhooyr.io/websocket (v1.7.4)
library because the former is unmaintained by the Golang team. To avoid
issues in the future, it is better to use a better implementation of the
WebSocket standard.

This one has been chosen because it is lightweight and well maintained.