Add voice connection
Implemented the voice connection to a channel. For the moment, we only
connect to a fixed channel.

We added a function to handle the heartbeat of the voice connection.
Enventually, voice connections should be an independent type and this
method should be part of that type.

We also added a voice properties to the Bot type, same as before, this
properties should be part of a Voice type, to allow multiple voice
connections.

The current code handling the WebSocket and UDP connections should be
encapsulated into the Voice type methods.

As a reference this commit is a basic implementation of this part of the
discord documentation :

(https://discordapp.com/developers/docs/topics/voice-connections)