Automatically select voice channel and code format
Some code has been implemented in order to retrieve the channel the bot
should join. It is looking in which channel the user sending the command
is. If the user is not in any channel, nothing happens.

The code has been reformated in order to move voice connections away
form the Bot type. We now have a Voice type, which can self contain a
voice connection. This allows the bot to be connected to one voice
channel per server.

The code for connecting to the voice servers (WebSocket and UDP
connections) has been moved into methods attached to the Voice type.