Handle voice disconnection and refactored code
This commit adds a channel than is able to signal voice disconnection to
the goroutines handling the voice connection operations (Reads and
Heartbeats). This is not handled perfectly and should be enhanced.

This commit also includes a reformating of the Bot type event handler. A
separate function as been created for each event type.

The code for the Voice type as been regrouped.